Rework developer API for this library
We introduce the concept of a View which now there are two. Once is the
list of tasks, the other is a view for a single task. This commit
introduces an IFrame manager class which gives us a standard API for
launching every view. We are trying to keep the developer experience
similar across all views. In the near future, we want to support
granular UI configurations such as themeing and controlling the display
of certain elements.