Make checkAuth work if previous responses failed
If a previous checkAuth failed because `onGetAuthToken` returned a bad
response, this Auth result would get memoized for all time. This commit
updates the logic so that if the result doesn't return a `token`, it
will make another request. It does some book-keeping as `checkAuth` may
be called in multiple places simultaneously and we only want one of the
places to make the request. The remainder can await the new promise.