Support recording before onAuthTokenRequested is known
We want to allow recordings to start before an end-user is logged in.
Once logged in, the recorder can have `onAuthTokenRequested` set and
then it can send the events that it has stored up to the ComplyCo
system.

This commit restructures EventSaver making a bunch of properties
private, allowing control of the `fetch` method so that we can test it
and adding a url validation method to ensure data is only getting sent
to ComplyCo via encrypted means.

This commit also adds tests for EventSaver.